==> exp2/scripts/03_bin-lratio.R <==
## ** Bin up the log ratio eye data for plotting **
options(crayon.enabled = FALSE, tidyverse.quiet = TRUE)
suppressPackageStartupMessages(library("dplyr"))
library("tidyr")

bin_up <- function(x, full = FALSE) {
  agg <- x %>%
    group_by(Speaker, Listener, bin) %>%
    summarize(comp = sum(Competitor),
              targ = sum(Target), .groups = "drop") %>%
    mutate(lratio = log((targ + .5) / (comp + .5))) %>%
    arrange(Speaker, Listener, bin)

  if (full) agg else agg$lratio    
}

## compute the interaction effect
## so that we can later calculate the standard error
calc_interaction <- function(x) {
  x %>%
    mutate(cond = paste0(Speaker, "_", Listener)) %>%
    select(cond, bin, lratio) %>%
    spread(cond, lratio) %>%
    mutate(eff = (early_early - early_late) -
             (late_early - late_late)) %>%
    `[[`("eff")
}

boot_once <- function(x) {
  nest(x, data = c(-SessionID)) %>%
    slice_sample(n = nrow(.), replace = TRUE) %>%
    unnest(data)
}

pog_subj_bin <- readRDS("data-derived/pog-window-subj.rds") %>%
  filter(Image %in% c("Target", "Competitor")) %>%
  mutate(bin = floor((frame_adj + 6) / 12) * 48) %>%
  group_by(SessionID, Speaker, Listener, bin, Image) %>%
  summarize(Y = sum(Y), .groups = "drop") %>%
  spread(Image, Y) %>%
  arrange(SessionID, Speaker, Listener, bin)

orig <- bin_up(pog_subj_bin, TRUE)

ix_eff <- replicate(1000,
                    pog_subj_bin %>% 
                    boot_once() %>%
                    bin_up(TRUE) %>%
                    calc_interaction())

sderr <- tibble(bin = orig$bin %>% unique() %>% sort(),
                se = apply(ix_eff, 1, sd))

orig2 <- orig %>%
  inner_join(sderr, "bin") %>%
  mutate(lmin = lratio - se,
         lmax = lratio + se) %>%
  select(-se, -comp, -targ)

saveRDS(orig2, "data-derived/pog-lratio-binned.rds")
